requirements.txt
certifi==2020.12.5
grpcio==1.36.1
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httpcore==0.12.3
httpx==0.17.0
hyperframe==5.2.0
idna==3.1
protobuf==3.15.5
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
